sphinx-agent: Move regexp patterns to `sphinx_agent.types`
This allows them to be reused in both the server and sphinx agent,
without introducing any import issues
