+++ lib/esbonio/esbonio/server/__init__.py
@@ -1,16 +1,12 @@
 from ._log import LOG_NAMESPACE
 from ._log import MemoryHandler
-from ._patterns import DIRECTIVE
-from ._patterns import ROLE
 from ._uri import Uri
 from .feature import LanguageFeature
 from .server import EsbonioLanguageServer
 from .server import EsbonioWorkspace
 
 __all__ = [
-    "DIRECTIVE",
     "LOG_NAMESPACE",
-    "ROLE",
     "EsbonioLanguageServer",
     "EsbonioWorkspace",
     "LanguageFeature",
